Add AlterShardingTableRuleBackendHandler
